chore: move new ops package to operations2 package at root
