+++ .mockery.yml
@@ -99,6 +99,15 @@ packages:
       structname: "{{.Mock}}{{.InterfaceName}}"
     interfaces:
       OnchainClient:
+  github.com/smartcontractkit/chainlink-deployments-framework/chain/evm/operations2/contract:
+    config:
+      all: false
+      pkgname: "mocks"
+      dir: "chain/evm/operations2/contract/mocks"
+      filename: "mock_{{.InterfaceName | snakecase}}.go"
+      structname: "{{.Mock}}{{.InterfaceName}}"
+    interfaces:
+      AccessControlContract:
   github.com/smartcontractkit/chainlink-deployments-framework/experimental/analyzer:
     config:
       all: false
