+++ chain/evm/operations2/contract/function_test.go
