+++ chain/evm/operations2/contract/read.go
