+++ chain/evm/operations2/contract/read_test.go
