--- tools/operations-gen/.mockery.yml
@@ -1,22 +1,0 @@
-all: false
-dir: "{{.InterfaceDir}}"
-filename: "mocks_test.go"
-force-file-write: true
-formatter: goimports
-log-level: info
-structname: "{{.Mock}}{{.InterfaceName}}"
-pkgname: "{{.SrcPackageName}}"
-recursive: false
-require-template-schema-exists: true
-template: testify
-template-schema: "{{.Template}}.schema.json"
-packages:
-  github.com/smartcontractkit/chainlink-deployments-framework/tools/operations-gen/contract:
-    config:
-      all: false
-      pkgname: "mocks"
-      dir: "contract/mocks"
-      filename: "mock_{{.InterfaceName | snakecase}}.go"
-      structname: "{{.Mock}}{{.InterfaceName}}"
-    interfaces:
-      AccessControlContract:
