+++ tools/operations-gen/generate/templates/evm/operations.tmpl
@@ -18,7 +18,7 @@ import (
 {{- if .Operations}}
 	cldf_evm "github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
 {{- end}}
-	"github.com/smartcontractkit/chainlink-deployments-framework/tools/operations-gen/contract"
+	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm/operations2/contract"
 {{- if .Operations}}
 	cld_ops "github.com/smartcontractkit/chainlink-deployments-framework/operations"
 {{- end}}
